Added Comment text validation and moved CommentForm Type inside Form folder
